requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.6.0
astroid==2.15.6
asttokens==2.4.1
async-timeout==4.0.3
attrs==22.2.0
bandit==1.7.5
black==23.7.0
boltons==21.0.0
bracex==2.3.post1
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.5
click-option-group==0.5.6
colorama==0.4.6
coverage==7.2.7
crashtest==0.4.1
cryptography==41.0.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.1.7
django-allauth==0.54.0
django-cors-headers==4.3.0
django-extensions==3.2.3
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dodgy==0.2.1
dulwich==0.21.3
exceptiongroup==1.1.2
executing==2.0.1
face==22.0.0
filelock==3.10.7
flake8==5.0.4
flake8-polyfill==1.0.2
frozenlist==1.4.1
geoip2==4.8.0
gitdb==4.0.10
GitPython==3.1.32
glom==22.1.0
html5lib==1.1
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
installer==0.7.0
ipython==8.20.0
isort==5.12.0
jaraco.classes==3.2.3
jedi==0.19.1
jsonschema==4.17.3
keyring==23.13.1
lazy-object-proxy==1.9.0
lockfile==0.12.2
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
maxminddb==2.5.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
multidict==6.0.4
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pbr==5.11.1
peewee==3.16.2
pep8-naming==0.10.0
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.2.0
poetry==1.4.2
poetry-core==1.5.2
poetry-plugin-export==1.3.0
prompt-toolkit==3.0.43
prospector==1.10.2
psycopg2==2.9.7
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pydocstyle==6.3.0
pyflakes==2.5.0
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyotp==2.9.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
pytest-mock==3.11.1
python-lsp-jsonrpc==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==2.13.7
requests==2.28.2
requests-mock==1.11.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
requirements-detector==1.2.2
rich==13.4.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semgrep==1.34.1
semver==3.0.1
setoptconf-tmp==0.3.1
shellingham==1.5.0.post1
six==1.16.0
smmap==5.0.0
snowballstemmer==2.2.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
stack-data==0.6.3
stevedore==5.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.7
traitlets==5.14.1
trove-classifiers==2023.3.9
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.7.1
ua-parser==0.18.0
ujson==5.8.0
urllib3==1.26.15
user-agents==2.2.0
virtualenv==20.21.0
wcmatch==8.4.1
wcwidth==0.2.13
webencodings==0.5.1
whitenoise==6.6.0
wrapt==1.15.0
xattr==0.10.1
yarl==1.9.4
zipp==3.15.0
